Move stuff used just in BaselineChecker into that file
